Split the large bdx module into sub-modules
