+++ rs-matter/Cargo.toml
@@ -135,6 +135,7 @@ large-buffers = [] # TCP support
 rs-matter-macros = { version = "0.1", path = "../rs-matter-macros" }
 cfg-if = "1"
 bitflags = { version =  "2.5", default-features = false }
+embedded-io-async = "0.7"
 heapless = "0.9"
 num = { version = "0.4", default-features = false }
 num-derive = "0.4"
